+++ lib/league_statistics.rb
@@ -1,0 +1,36 @@
+require 'csv'
+require_relative './stat_tracker'
+require_relative './game'
+require_relative './team'
+require_relative './game_team'
+require_relative './season_statistics'
+
+class LeagueStats < SeasonStats
+   def count_of_teams
+      @data_teams.count  
+   end
+
+   def best_offense
+      convert_team_id_to_name(highest_average_team_id(team_stats))
+   end
+
+   def worst_offense
+      convert_team_id_to_name(lowest_average_team_id(team_stats))
+   end
+
+   def highest_scoring_visitor
+      convert_team_id_to_name(highest_average_team_id(team_stats_hoa("away")))
+   end
+
+   def highest_scoring_home_team
+      convert_team_id_to_name(highest_average_team_id(team_stats_hoa("home")))
+   end
+
+   def lowest_scoring_visitor
+      convert_team_id_to_name(lowest_average_team_id(team_stats_hoa("away")))
+   end
+
+   def lowest_scoring_home_team
+      convert_team_id_to_name(lowest_average_team_id(team_stats_hoa("home")))
+   end
+end
